third_party/blink/web_tests/external/wpt/html/interaction/focus/processing-model/focus-fixup-rule-one-no-dialogs-expected.txt

This is a testharness.js-based test.
[FAIL] #button1
  assert_equals: active element isn't fixed-up sync expected Element node <button id="button1" disabled="">Button 1</button> but got Element node <body><div>\n  <button id="button1" disabled="">Button 1</...
[FAIL] #button2
  assert_equals: active element isn't fixed-up sync expected Element node <button id="button2" hidden="">Button 2</button> but got Element node <body><div>\n  <button id="button1" disabled="">Button 1</...
[FAIL] #button4
  assert_equals: active element isn't fixed-up sync expected Element node <button id="button4">Button 4</button> but got Element node <body><div>\n  <button id="button1" disabled="">Button 1</...
[FAIL] #button5
  assert_equals: active element isn't fixed-up sync expected Element node <button id="button5">Button 5</button> but got Element node <body><div>\n  <button id="button1" disabled="">Button 1</...
[FAIL] #div
  assert_equals: active element isn't fixed-up sync expected Element node <div id="div">Div</div> but got Element node <body><div>\n  <button id="button1" disabled="">Button 1</...
[FAIL] #editable
  assert_equals: active element isn't fixed-up sync expected Element node <div id="editable" contenteditable="false">editor</div> but got Element node <body><div>\n  <button id="button1" disabled="">Button 1</...
Harness: the test ran to completion.
